+++ webapp/app/forms/flows/lotse_flow.py
@@ -25,8 +25,8 @@ from app.forms.steps.lotse.steuerminderungen import StepVorsorge, StepAussergBel
 from app.forms.steps.lotse_multistep_flow_steps.confirmation_steps import StepConfirmation, StepAck, StepFiling
 from app.forms.steps.lotse_multistep_flow_steps.declaration_steps import StepDeclarationIncomes, StepDeclarationEdaten, StepSessionNote
 from app.forms.steps.lotse.personal_data import StepSteuernummer, StepPersonB, StepTelephoneNumber, StepPersonA, \
-    StepPersonAHasDisability, StepPersonBHasDisability, StepPersonAHasDisability, StepPersonBHasDisability, \
-    StepPauschbetragPersonA, StepPauschbetragPersonB
+    StepPersonAHasDisability, StepPersonBHasDisability
+from app.forms.steps.lotse.pauschbetrag import StepPauschbetragPersonA, StepPauschbetragPersonB
 from app.forms.steps.lotse_multistep_flow_steps.personal_data_steps import StepIban, StepFamilienstand
 from app.forms.steps.step import Section
 from app.model.form_data import MandatoryFormData, MandatoryConfirmations, \
